Make State::max_chain_length u16

+++ zlib-rs/src/deflate/longest_match.rs
@@ -26,7 +26,7 @@ fn longest_match_help<const SLOW: bool>(
     let limit_base: Pos;
     let early_exit: bool;
 
-    let mut chain_length: usize;
+    let mut chain_length: u16;
     let mut best_len: usize;
 
     let lookahead = state.lookahead;
